Remove autocomplete mode for now
Although explicitly disabling it in python mode, it is getting
activated with the current settings, and I couldn't find out why. It's
slowing down editing extremely, so I'll disable it for now, until I
have the time to check and maybe rewrite the initialization of how and
when to use autocomplete or company mode.